refactor: repositories model list
