+++ src/pages/products/[product]/repositories/[repository]/components/GithubRepositoriesModal/index.ts
@@ -1,0 +1,1 @@
+export { GithubRepositoriesModal } from './GithubRepositoriesModal';
